Trying to wire it all together.
host/stay-fresh-listener.js
---------------------------
Listens for HTTP requests on port 7700. The only route configured is
GET /:eventName, which is supposed to send a message on process.stdout with
the value of :eventName. The Chrome extension should be listening for stuff
sent on stdout.

main-background.js
------------------
The script run in the generated "event page" of the extension. Formerly
known as the "background page." The script basically just has some two
listeners set up. One for the sendMessage events that popup.html will send
it ("user clicked start listening", "user clicked stop listening") and one
for the native messages that stay-fresh-listener.js sends it.

popup.html
----------
Has two buttons to start and stop listening and loads popup.js.

popup.js
--------
Responds to clicks on the "start" and "stop" buttons and sends messages to
the "event page" (i.e., the "background page") which is supposed to ad the
relevent tab id to it's list of "listeningTabIds" which need to be reloaded
on certain Vim events.