Keeping it simple
No longer creating a popup.html. Adding a toggle button to that was turning
out to be a giant black hole. Also the reloader doesn't care about the URL
anymore, only the tab id. If the URL changes within a listening tab, it will
keep auto-reloading until user deactivates. In practice this makes more
sense anyway.