Fix default_icon bug for Retina displays
Due to the logic described in the link below, Chrome was assuming that
default_icon was a 19-pixel-square icon and thus unsuitable for Retina
displays, which need a 38-pixel-square icon to look best. Chrome had logic
then to pull in one of the larger icons specified in the "icons" section.
Changed the manifest to expllicitly declare a usable 38-pixel-square icon.

https://code.google.com/p/chromium/issues/detail?id=404572&q=default_icon&colspec=ID%20Pri%20M%20Iteration%20ReleaseBlock%20Cr%20Status%20Owner%20Summary%20OS%20Modified